+++ sportex.api.web/DTO/LocationDTO.cs
@@ -12,8 +12,8 @@ namespace sportex.api.web.DTO
         public int ID { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int? Latitude { get; set; }
-        public int? Longitude { get; set; }
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; }
         #endregion
 
         public LocationDTO()
